+++ Rakefile
@@ -23,7 +23,7 @@ end
 
 desc "Generate the API client files based on the OpenAPI route docs."
 task :generate do
-  require_relative "lib/open_api_client_generator"
+  require_relative "lib/openapi_client_generator"
   OpenAPIClientGenerator::API.at(OasParser::Definition.resolve("../routes/openapi/api.github.com/index.json")) do |api|
     File.open("lib/octokit/client/#{api.resource}.rb", "w") do |f|
       f.puts api.to_s
