+++ lib/openapi_client_generator.rb
